Full PIO SPI Implementation
Add PIO-based SPI slave transport layer (Mode 0) and full project scaffolding

This commit introduces the first hardware-usable serial transport layer for
the Modular Control Module (MCM), implementing a PIO-based SPI slave designed
for deterministic, low-latency control data exchange with a master MCU.

Key additions and architectural decisions:

- PIO-based SPI slave (Mode 0, MSB-first):
  Introduces a dedicated PIO program for SPI slave operation using CS-framed,
  fixed-length transactions. The design prioritizes robustness and clarity
  over throughput, with one 8-byte packet per CS assertion.

- Fixed-length, versioned binary protocol:
  Establishes an 8-byte wire protocol for authoritative parameter state
  publishing, including sync byte, protocol versioning, message type,
  parameter index, 24-bit value payload, and CRC-8 protection.

- Transport abstraction layer:
  Adds a TransportSPI class that cleanly separates SPI signaling from packet
  semantics, allowing encoder, parameter, and application logic to remain
  transport-agnostic.

- State publishing pipeline:
  Integrates the SPI transport with the existing event-driven architecture via
  a StatePublisher, converting internal parameter-change events into wire-ready
  packets with back-pressure handling.

- Full project layout and module separation:
  Formalizes the firmware structure into clear subsystems (PIO encoders,
  debounced buttons, parameter binding, event queue, protocol, and transport),
  establishing a scalable foundation for future features.

- Placeholder SPI pin assignments:
  SPI GPIO pins are intentionally left as explicit placeholders to allow final
  hardware pin mapping decisions without requiring architectural changes.

This commit completes the transition from architectural prototype to a
hardware-facing control-surface firmware. Future commits will finalize SPI pin
assignment, add master-to-slave command handling, parameter snapshot support,
and optional I2C and standalone MIDI/USB operation.

+++ src/TransportProtocol.h
@@ -1,4 +1,3 @@
-
 #pragma once
 #include <Arduino.h>
 
@@ -17,7 +16,7 @@ static inline uint8_t proto_crc8(const uint8_t* data, uint8_t len)
   for (uint8_t i = 0; i < len; i++) {
     crc ^= data[i];
     for (uint8_t b = 0; b < 8; b++) {
-      crc = (crc & 0x80) ? (crc << 1) ^ 0x07 : (crc << 1);
+      crc = (crc & 0x80) ? (uint8_t)((crc << 1) ^ 0x07) : (uint8_t)(crc << 1);
     }
   }
   return crc;
